Multimodal/requirements.txt
torch
torchvision
torchaudio
torchinfo
torchview
graphviz
numpy
pandas
tqdm
Pillow
matplotlib
seaborn
facenet-pytorch
scikit-learn
scipy
timm